requirements.txt
openai
python-dotenv
google-genai
firebase-admin
flask
flask-cors
gunicorn